requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google AI (Gemini)
google-generativeai==0.3.1

# Vector Database
qdrant-client>=1.7.1

# Embeddings
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.0

# Bio libraries
biopython==1.81

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0

# Database (Optional)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
